fix relationship cart and cart item entity

+++ src/main/java/com/example/backend_demo/model/Order.java
@@ -40,7 +40,7 @@ public class Order {
     @Embedded
     private PaymentDetails paymentDetails = new PaymentDetails();
 
-    private double totalPrice;
+    private int totalPrice;
 
     private Integer totalDiscountedPrice;
